feat: ADR-0030 money Decimal + WeChat sub-package split (#34)
* feat(money): migrate Float -> Decimal(10,2) end-to-end (ADR-0030)

- Models: Order.price / Payment.amount -> Numeric(10, 2), typed Decimal
- Services: SERVICE_PRICES + PaymentService + WalletService + admin refund
  ratio computed in Decimal with quantize(ROUND_HALF_UP)
- Provider DTO: amount_yuan / total_yuan / refund_yuan now Decimal,
  with _to_decimal() helper accepting Decimal|int|float|str
- Schema: Pydantic field_serializer keeps API output as JSON number
  (e.g. 299.0) so frontend contract is unchanged
- Alembic: a1d0c0de0030 Float->Numeric(10,2) (PG USING + SQLite batch)
- WeChat: formatPrice already number/string tolerant; added ADR-0030 anchor
- iOS: Order.price / Payment.amount already Decimal
- Tests: +13 boundary cases in tests/unit/test_decimal_money.py
  Full backend regression: 880 passed / 15 skipped / 108s
- Docs: ADR-0030 + DECISION_LOG D-041 + TECH_DEBT TD-ARCH-03 closed

Closes TD-ARCH-03.

* fix(emergency): add description to 5 endpoints to satisfy api-docs-check

* feat(wechat): split into 19 sub-packages + lift 4 global components

- app.json: 11 main pages + 19 subPackages (all 31 pages covered)
- preloadRule for hot paths (patient home, companion home, profile)
- Lift loading-overlay/empty-state/patient-tab-bar/companion-tab-bar to global
- Strip 14 page json files of duplicated component declarations
- Add components/README.md catalog (4 categories, 9 components)
- Add app-json structure tests (8 cases): page count, subpkg unique, preload refs
- Regenerate docs/api/* after emergency endpoint description fixes

Tests: 221 + 8 new = 229 passed (jest)

---------

+++ backend/app/models/payment.py
@@ -1,7 +1,8 @@
 import uuid
 from datetime import datetime, timezone
+from decimal import Decimal
 
-from sqlalchemy import DateTime, Float, String, Text, UniqueConstraint, Uuid
+from sqlalchemy import DateTime, Numeric, String, Text, UniqueConstraint, Uuid
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.database import Base
@@ -22,7 +23,8 @@ class Payment(Base):
     user_id: Mapped[uuid.UUID] = mapped_column(
         Uuid(as_uuid=True), nullable=False
     )
-    amount: Mapped[float] = mapped_column(Float, nullable=False)
+    # ADR-0030: 金额使用 Decimal/Numeric(10,2)
+    amount: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
     payment_type: Mapped[str] = mapped_column(
         String(20), nullable=False
     )  # "pay" or "refund"
